reorganize project
using this principle:

- widgets only used in one screen are put
  in the `*_screen` folder
- widgets which are shared are put in
  the `widgets` folder
