+++ lib/dashboard_screen.dart
@@ -1,4 +1,4 @@
-import 'package:dream_home/dashboard/gauge.dart';
+import 'package:dream_home/dashboard_screen/gauge.dart';
 import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
