+++ lib/main.dart
@@ -1,4 +1,4 @@
-import 'package:dream_home/dashboard.dart';
+import 'package:dream_home/dashboard_screen.dart';
 import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
